My pipelined version - works 3x faster!
It finishes like this:

    Received 45/45
    tb/mandel_tb.vhdl:164:9:@7780ns:(assertion note):
        Successful end of test

So it takes 7.78us to process our 45 inputs.
Compare that to the serially-processing algorithm:

    tb/mandel_tb.vhdl:172:5:@22170ns:(assertion note):
        Successful end of test

...taking 22.17us for the same inputs!

Pipelining was hard to do - but it really delivers.
Basically, the 3 multipliers are almost constantly
kept busy, they never stop unless the pipeline is empty.
As long as we keep feeding the engine with inputs,
they are doing work on every cycle - as opposed to the
serial version, that only uses them in the first stages.

Excellent - now it's time to run this on my Spartan3... :-)